fix(explore): render the cover strip on its own, and only on the webapp
Review of #6613 found the Explore strip was borrowing the new tab's
whole state machine, and paying for it four ways:

- The extension new tab renders MainFeedLayout too, and its in-place
  Explore tab resolves to OtherFeedPage.Explore, so the Explore strip
  rendered inside the extension — stacked on the extension's own strip
  on layout v1, and opening auth inline where the extension origin
  cannot run OAuth. It is now webapp-only.
- HijackingHeroStrip short-circuits to the "Welcome back" panel when a
  signBack record exists, and that record survives logout. A visitor
  who once logged in saw their name and email on a public, indexed
  page. The Explore surface now renders the cover strip's visual
  directly, with its own copy and its own inline auth.
- Every impression carried the experiment's target_id. The Explore
  strip logs under its own.
- Importing the experiment module pulled the registration forms into
  the shared mainFeedLayout chunk and the Squad directory. Only the
  cover strip's visual is shared now (`HijackingCoverStrip`); the
  experiment moves back to the extension, where its single consumer is.

Two smaller things the review caught: the strip rendered null until
boot answered, so it was spliced in above the H1 on statically
generated pages — it now holds its slot with a placeholder of the same
height; and the Squad directory's section gained an unconditional
gap-6 that spaced the loading skeleton for logged-in members who never
see the strip — the strip now sits between the header and the section
and carries its own margin.

The Explore hub's feed list moves into useFeedName beside the other
feed lists, and sources/[source] uses its own inset constant.
